// client/src/interfaces/RegisterFormValues.ts
export interface RegisterFormValues {
    first: string;
    middle: string;
    last: string;
    phone: string;
    email: string;
    password: string;
    image: string;
    alt: string;
    state: string;
    country: string;
    city: string;
    street: string;
    houseNumber: number;
    zip: number | null;
    isBusiness: boolean;
  }